ogs-sha[12]-hmac: Ensure input (key, message) are not modified
When calling a HMAC API, the input is the key and the message,
and the output is the generate HMAC.  Functions should not modify
the input data, as this can lead to all kinds of strange side-effects.

Let's ensure input data are not modified and subsequently mark input
arguments with 'const' qualifier.